+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 1.18.5
 ======
 
+* bug-fix: pass kms id as parameter for uploading code with Server side encryption
 * feature: ``PipelineModel``: Create a Transformer from a PipelineModel
 * bug-fix: ``AlgorithmEstimator``: Make SupportedHyperParameters optional
 * feature: ``Hyperparameter``: Support scaling hyperparameters